#107 [refactor] 서버 도메인으로 변경
