+++ app/src/main/java/com/sopt/peekabookaos/data/repository/DetailRepositoryImpl.kt
@@ -1,16 +1,17 @@
 package com.sopt.peekabookaos.data.repository
 
 import com.sopt.peekabookaos.data.entity.NoResponse
-import com.sopt.peekabookaos.data.entity.response.DetailResponse
 import com.sopt.peekabookaos.data.source.remote.DetailDataSource
+import com.sopt.peekabookaos.domain.entity.Detail
+import com.sopt.peekabookaos.domain.repository.DetailRepository
 import javax.inject.Inject
 
 class DetailRepositoryImpl @Inject constructor(
     private val detailDataSource: DetailDataSource
 ) : DetailRepository {
-    override suspend fun getDetail(bookId: Int): Result<DetailResponse> =
+    override suspend fun getDetail(bookId: Int): Result<Detail> =
         kotlin.runCatching { detailDataSource.getDetail(bookId) }.map { response ->
-            response.data!!
+            response.data!!.detail()
         }
 
     override suspend fun deleteDetail(bookId: Int): Result<NoResponse> =
