+++ app/src/main/java/com/sopt/peekabookaos/di/RepositoryModule.kt
@@ -2,7 +2,6 @@ package com.sopt.peekabookaos.di
 
 import com.sopt.peekabookaos.data.repository.CreateUpdateRepository
 import com.sopt.peekabookaos.data.repository.CreateUpdateRepositoryImpl
-import com.sopt.peekabookaos.data.repository.DetailRepository
 import com.sopt.peekabookaos.data.repository.DetailRepositoryImpl
 import com.sopt.peekabookaos.data.repository.NaverRepository
 import com.sopt.peekabookaos.data.repository.NaverRepositoryImpl
@@ -13,6 +12,7 @@ import com.sopt.peekabookaos.data.repository.RecommendRepositoryImpl
 import com.sopt.peekabookaos.data.repository.SearchRepositoryImpl
 import com.sopt.peekabookaos.data.repository.ShelfRepository
 import com.sopt.peekabookaos.data.repository.ShelfRepositoryImpl
+import com.sopt.peekabookaos.domain.repository.DetailRepository
 import com.sopt.peekabookaos.domain.repository.SearchRepository
 import dagger.Binds
 import dagger.Module
