+++ app/src/main/java/com/sopt/peekabookaos/domain/repository/DetailRepository.kt
@@ -1,0 +1,10 @@
+package com.sopt.peekabookaos.domain.repository
+
+import com.sopt.peekabookaos.data.entity.NoResponse
+import com.sopt.peekabookaos.domain.entity.Detail
+
+interface DetailRepository {
+    suspend fun getDetail(bookId: Int): Result<Detail>
+
+    suspend fun deleteDetail(bookId: Int): Result<NoResponse>
+}
